Use freetype APIs from dependency crates as much as possible.
