+++ engine/Cargo.toml
@@ -31,12 +31,12 @@ libc = "0.2"
 chrono = "0.4.9"
 #log = { version = "0.4", features = ["std"] }
 harfbuzz-sys = {version = "^0.3", default-features = false}
+freetype-rs = "0.23.0"
 
 [target.'cfg(target_os = "macos")'.dependencies]
 core-foundation = "0.7.0"
 core-graphics = "0.19.0"
 core-text = "15.0.0"
-freetype-rs = "0.23.0"
 objc = "0.2.7"
 objc-foundation = "0.1.1"
 objc_id = "0.1.1"
